added logger

+++ src/hooks/timeStampToDate.js
@@ -1,7 +1,7 @@
 export const timeStampToDate = (timeStamp)=>{
     if(timeStamp){
         return Intl.DateTimeFormat('fa', {
-            useGrouping: false, year: "numeric", month: "long", day: "numeric"
+            useGrouping: false, year: "numeric", month: "numeric", day: "numeric"
         }).format(timeStamp * 1000);
     }
 }
